Fixed mobile view, broke apart Resources into components
