+++ site/src/app/views/resources/Resources.jsx
@@ -1,19 +1,8 @@
-import { useState, useEffect } from "react";
-import { Tooltip, Footer } from "src/app/components";
-import { starterPackData } from "src/app/data/resources-info";
-import { Carousel, Stack, Card } from "react-bootstrap";
-import { Nav, Redirect } from "src/app/components";
+import { Nav } from "src/app/components";
+import ResourceCarousel from "src/app/components/resource-carousel/ResourceCarousel";
 import useResources from "./useResources";
 
 import "./Resources.scss";
-
-function getWindowDimensions() {
-	const { innerWidth: width, innerHeight: height } = window;
-	return {
-		width,
-		height,
-	};
-}
 
 const backgroundImages = [
 	"/assets/images/resources-card1.png",
@@ -35,9 +24,7 @@ export default function Resources() {
 			<div id="right-fishes" />
 
 			<center>
-				<h1 style={{ fontSize: "100px", textShadow: "#68B3D7 10px 10px 10px" }}>
-					RESOURCES
-				</h1>
+				<h1 className="resources-title">RESOURCES</h1>
 			</center>
 			<img
 				src="/assets/images/resources-coral-top.svg"
@@ -46,115 +33,21 @@ export default function Resources() {
 			/>
 			<Nav />
 			<div className="resource-list">
-				{resources.order.map(
-					({ _id, iconUrl, title, description, resources }, index) => (
-						<div key={_id} className="category">
-							<h3 style={{ fontSize: "100px" }}>{title}</h3>
-							<ResourceCarousel
-								resources={resources}
-								backgroundImage={
-									backgroundImages[
-										Math.floor(
-											(index / resourcesLength) * backgroundImages.length
-										)
-									]
-								}
-								className="resource-carousel"
-							/>
-						</div>
-					)
-				)}
-			</div>
-		</div>
-	);
-}
-
-function ResourceCarousel({ resources, backgroundImage }) {
-	const [currentIndex, setCurrentIndex] = useState(0);
-	const [isTransitioning, setIsTransitioning] = useState(true);
-	const totalSlides = resources.length;
-
-	const nextResource = () => {
-		setCurrentIndex((prevIndex) => (prevIndex + 1) % resources.length);
-	};
-
-	const prevResource = () => {
-		setCurrentIndex(
-			(prevIndex) => (prevIndex - 1 + resources.length) % resources.length
-		);
-	};
-
-	// Transition back to first card
-	useEffect(() => {
-		if (currentIndex === totalSlides) {
-			setTimeout(() => {
-				setIsTransitioning(false);
-				setCurrentIndex(0);
-			}, 400);
-		}
-	}, [currentIndex, totalSlides]);
-
-	return (
-		<div
-			style={{ display: "flex", flexDirection: "column", alignItems: "center" }}
-		>
-			{/* Carousel Section */}
-			<div className="carousel-container">
-				<button
-					className="carousel-button left-button"
-					onClick={prevResource}
-				></button>
-
-				<div
-					className="carousel-wrapper container"
-					style={{
-						backgroundImage: `url(${backgroundImage})`,
-					}}
-				>
-					<div
-						className="carousel-track"
-						style={{
-							transform: `translateX(-${currentIndex * 100}%)`,
-							transition: "transform 0.4s ease-in-out",
-						}}
-					>
-						{resources.map(
-							({ _id, resourceIconUrl, title, link, description }) => (
-								<div key={_id} className="resource-card container">
-									<a
-										href={link}
-										target="_blank"
-										rel="noopener noreferrer"
-										className="card-link"
-									>
-										<img
-											src={resourceIconUrl}
-											alt={title}
-											className="resource-icon"
-										/>
-										<h4 className="resource-title">{title}</h4>
-										<p className="resource-description">{description}</p>
-									</a>
-								</div>
-							)
-						)}
+				{resources.order.map(({ _id, title, resources }, index) => (
+					<div key={_id} className="category">
+						<h3 className="category-title">{title}</h3>
+						<ResourceCarousel
+							resources={resources}
+							backgroundImage={
+								backgroundImages[
+									Math.floor(
+										(index / resourcesLength) * backgroundImages.length
+									)
+								]
+							}
+							className="resource-carousel"
+						/>
 					</div>
-				</div>
-
-				<button
-					className="carousel-button right-button"
-					onClick={nextResource}
-				></button>
-			</div>
-
-			{/* Indicators Section (now outside and below the carousel) */}
-			<div className="carousel-indicators">
-				{resources.map((_, index) => (
-					<span
-						key={index}
-						className={`dot ${index === currentIndex ? "active" : ""}`}
-						onClick={() => setCurrentIndex(index)}
-					/>
 				))}
 			</div>
 		</div>
